+++ tgbotapi.utils/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/utils/extensions/raw/PreCheckoutQuery.kt
@@ -1,7 +1,7 @@
 package dev.inmo.tgbotapi.extensions.utils.extensions.raw
 
 import dev.inmo.tgbotapi.types.InvoicePayload
-import dev.inmo.tgbotapi.types.ShippingOptionIdentifier
+import dev.inmo.tgbotapi.types.ShippingOptionId
 import dev.inmo.tgbotapi.types.payments.OrderInfo
 import dev.inmo.tgbotapi.types.payments.PreCheckoutQuery
 import dev.inmo.tgbotapi.utils.RiskFeature
@@ -13,7 +13,7 @@ val PreCheckoutQuery.total_amount: Long
 val PreCheckoutQuery.invoice_payload: InvoicePayload
     get() = invoicePayload
 @RiskFeature(RawFieldsUsageWarning)
-val PreCheckoutQuery.shipping_option_id: ShippingOptionIdentifier?
+val PreCheckoutQuery.shipping_option_id: ShippingOptionId?
     get() = shippingOptionId
 @RiskFeature(RawFieldsUsageWarning)
 val PreCheckoutQuery.order_info: OrderInfo?